+++ changelog.md
@@ -3,3 +3,7 @@
 - [rcamera.c] RENAMED: `Init_Camera` to `InitDefaultCamera`
 - [rcamera.c] MODIFIED: 若干变量改为宏。
 
+# V1.1.6
+- [rmodels.c] DELETED: `GenMeshPlane`。
+- [rmodels.c] DELETED: `GenMeshPoly`。
+- [rcamera.c] MOVED: `InitDefaultCamera` to `rcore.c`。